+++ assets/js/dashboard/stats/countries.js
@@ -8,11 +8,13 @@ import Bar from './bar'
 import MoreLink from './more-link'
 import * as api from '../api'
 import { navigateToQuery } from '../query'
-
+import { withThemeConsumer } from '../theme-consumer-hoc';
 class Countries extends React.Component {
   constructor(props) {
     super(props)
     this.resizeMap = this.resizeMap.bind(this)
+    this.drawMap = this.drawMap.bind(this)
+    this.getDataset = this.getDataset.bind(this)
     this.state = {loading: true}
   }
 
@@ -31,6 +33,14 @@ class Countries extends React.Component {
       this.setState({loading: true, countries: null})
       this.fetchCountries().then(this.drawMap.bind(this))
     }
+
+    if (this.props.darkTheme !== prevProps.darkTheme) {
+      if (document.getElementById('map-container')) {
+        document.getElementById('map-container').removeChild(document.querySelector('.datamaps-hoverover'));
+        document.getElementById('map-container').removeChild(document.querySelector('.datamap'));
+      }
+      this.drawMap();
+    }
   }
 
   getDataset() {
@@ -41,7 +51,10 @@ class Countries extends React.Component {
 
     var paletteScale = d3.scale.linear()
       .domain([0,maxValue])
-      .range(["#f3ebff","#a779e9"]);
+      .range([
+        this.props.darkTheme ? "#2e3954" : "#f3ebff",
+        this.props.darkTheme ? "#6366f1" : "#a779e9"
+      ]);
 
     this.state.countries.forEach(function(item){
       dataset[item.name] = {numberOfThings: item.count, fillColor: paletteScale(item.count)};
@@ -68,26 +81,30 @@ class Countries extends React.Component {
   drawMap() {
     var dataset = this.getDataset();
     const label = this.props.query.period === 'realtime' ? 'Current visitors' : 'Visitors'
+    const defaultFill = this.props.darkTheme ? '#2d3747' : '#f8fafc'
+    const highlightFill = this.props.darkTheme ? '#374151' : '#F5F5F5'
+    const borderColor = this.props.darkTheme ? '#1f2937' : '#dae1e7'
+    const highlightBorderColor = this.props.darkTheme ? '#4f46e5' : '#a779e9'
 
     this.map = new Datamap({
       element: document.getElementById('map-container'),
       responsive: true,
       projection: 'mercator',
-      fills: { defaultFill: '#f8fafc' },
+      fills: { defaultFill },
       data: dataset,
       geographyConfig: {
-        borderColor: '#dae1e7',
+        borderColor,
         highlightBorderWidth: 2,
         highlightFillColor: function(geo) {
-          return geo['fillColor'] || '#F5F5F5';
+          return geo['fillColor'] || highlightFill;
         },
-        highlightBorderColor: '#a779e9',
+        highlightBorderColor,
         popupTemplate: function(geo, data) {
           if (!data) { return ; }
           const pluralizedLabel = data.numberOfThings === 1 ? label.slice(0, -1) : label
-          return ['<div class="hoverinfo">',
+          return ['<div class="hoverinfo dark:bg-gray-800 dark:shadow-gray-850 dark:border-gray-850 dark:text-gray-200">',
             '<strong>', geo.properties.name, '</strong>',
-            '<br><strong>', numberFormatter(data.numberOfThings), '</strong> ' + pluralizedLabel,
+            '<br><strong class="dark:text-indigo-400">', numberFormatter(data.numberOfThings), '</strong> ' + pluralizedLabel,
             '</div>'].join('');
         }
       },
@@ -109,7 +126,7 @@ class Countries extends React.Component {
     if (this.state.countries) {
       return (
         <React.Fragment>
-          <h3 className="font-bold">Countries</h3>
+          <h3 className="font-bold dark:text-gray-100">Countries</h3>
           <div className="mt-6 mx-auto" style={{width: '100%', maxWidth: '475px', height: '320px'}} id="map-container"></div>
           <MoreLink site={this.props.site} list={this.state.countries} endpoint="countries" />
         </React.Fragment>
@@ -119,7 +136,7 @@ class Countries extends React.Component {
 
   render() {
     return (
-      <div className="stats-item relative bg-white shadow-xl rounded p-4" style={{height: '436px'}}>
+      <div className="stats-item relative bg-white dark:bg-gray-825 shadow-xl rounded p-4" style={{height: '436px'}}>
         { this.state.loading && <div className="loading my-32 mx-auto"><div></div></div> }
         <FadeIn show={!this.state.loading}>
           { this.renderBody() }
@@ -129,4 +146,4 @@ class Countries extends React.Component {
   }
 }
 
-export default withRouter(Countries)
+export default withRouter(withThemeConsumer(Countries))