+++ assets/js/dashboard/stats/sources/referrer-list.js
@@ -57,7 +57,7 @@ export default class Referrers extends React.Component {
     if (this.props.query.filters.source && this.props.query.filters.source !== 'Google' && referrer.name !== 'Direct / None') {
       return (
         <a target="_blank" href={'//' + referrer.name} className="hidden group-hover:block">
-          <svg className="inline h-4 w-4 ml-1 -mt-1 text-gray-600" fill="currentColor" viewBox="0 0 20 20"><path d="M11 3a1 1 0 100 2h2.586l-6.293 6.293a1 1 0 101.414 1.414L15 6.414V9a1 1 0 102 0V4a1 1 0 00-1-1h-5z"></path><path d="M5 5a2 2 0 00-2 2v8a2 2 0 002 2h8a2 2 0 002-2v-3a1 1 0 10-2 0v3H5V7h3a1 1 0 000-2H5z"></path></svg>
+          <svg className="inline h-4 w-4 ml-1 -mt-1 text-gray-600 dark:text-gray-400" fill="currentColor" viewBox="0 0 20 20"><path d="M11 3a1 1 0 100 2h2.586l-6.293 6.293a1 1 0 101.414 1.414L15 6.414V9a1 1 0 102 0V4a1 1 0 00-1-1h-5z"></path><path d="M5 5a2 2 0 00-2 2v8a2 2 0 002 2h8a2 2 0 002-2v-3a1 1 0 10-2 0v3H5V7h3a1 1 0 000-2H5z"></path></svg>
         </a>
       )
     }
@@ -71,16 +71,16 @@ export default class Referrers extends React.Component {
     return (
       <div className="flex items-center justify-between my-1 text-sm" key={referrer.name}>
         <div className="w-full h-8" style={{maxWidth: 'calc(100% - 4rem)'}}>
-          <Bar count={referrer.count} all={this.state.referrers} bg="bg-blue-50" />
+          <Bar count={referrer.count} all={this.state.referrers} bg="bg-blue-50 dark:bg-gray-500 dark:bg-opacity-15" />
           <span className="flex px-2 group" style={{marginTop: '-26px'}} >
-            <LinkOption className="block truncate" to={{search: query.toString()}} disabled={referrer.name === 'Direct / None'}>
+            <LinkOption className="block truncate dark:text-gray-300" to={{search: query.toString()}} disabled={referrer.name === 'Direct / None'}>
               <img src={`https://icons.duckduckgo.com/ip3/${referrer.url}.ico`} referrerPolicy="no-referrer" className="inline h-4 w-4 mr-2 align-middle -mt-px" />
               { referrer.name }
             </LinkOption>
             { this.renderExternalLink(referrer) }
           </span>
         </div>
-        <span className="font-medium">{numberFormatter(referrer.count)}</span>
+        <span className="font-medium dark:text-gray-200">{numberFormatter(referrer.count)}</span>
       </div>
     )
   }
@@ -93,7 +93,7 @@ export default class Referrers extends React.Component {
     if (this.state.referrers.length > 0) {
       return (
         <React.Fragment>
-          <div className="flex items-center mt-3 mb-2 justify-between text-gray-500 text-xs font-bold tracking-wide">
+          <div className="flex items-center mt-3 mb-2 justify-between text-gray-500 dark:text-gray-400 text-xs font-bold tracking-wide">
             <span>Referrer</span>
             <span>{ this.label() }</span>
           </div>
@@ -104,7 +104,7 @@ export default class Referrers extends React.Component {
         </React.Fragment>
       )
     } else {
-      return <div className="text-center mt-44 font-medium text-gray-500">No data yet</div>
+      return <div className="text-center mt-44 font-medium text-gray-500 dark:text-gray-400">No data yet</div>
     }
   }
 
@@ -112,7 +112,7 @@ export default class Referrers extends React.Component {
     if (this.state.referrers) {
       return (
         <React.Fragment>
-          <h3 className="font-bold">Top Referrers</h3>
+          <h3 className="font-bold dark:text-gray-100">Top Referrers</h3>
           { this.renderList() }
           <MoreLink site={this.props.site} list={this.state.referrers} endpoint={`referrers/${this.props.query.filters.source}`} />
         </React.Fragment>
@@ -122,7 +122,7 @@ export default class Referrers extends React.Component {
 
   render() {
     return (
-      <div className="stats-item relative bg-white shadow-xl rounded p-4" style={{height: '436px'}}>
+      <div className="stats-item relative bg-white dark:bg-gray-825 shadow-xl rounded p-4" style={{height: '436px'}}>
         { this.state.loading && <div className="loading mt-44 mx-auto"><div></div></div> }
         <FadeIn show={!this.state.loading}>
           { this.renderContent() }